+++ workspaces/marketplace/.changeset/good-colts-exist.md
@@ -1,5 +1,6 @@
 ---
 '@red-hat-developer-hub/backstage-plugin-marketplace': patch
+'@red-hat-developer-hub/backstage-plugin-marketplace-common': patch
 ---
 
 load all the package configurations
