+++ workspaces/marketplace/plugins/marketplace/src/components/MarketplacePackageInstallContent.tsx
@@ -19,21 +19,15 @@ import type { SetStateAction } from 'react';
 import { useCallback, useEffect, useState } from 'react';
 
 import { ErrorPage, Progress } from '@backstage/core-components';
-import {
-  alertApiRef,
-  useApi,
-  useRouteRef,
-  useRouteRefParams,
-} from '@backstage/core-plugin-api';
+import { useRouteRef, useRouteRefParams } from '@backstage/core-plugin-api';
 
 import yaml from 'yaml';
 import { useNavigate } from 'react-router-dom';
 
 import {
   MarketplacePackage,
-  MarketplacePackageAppConfigExamples,
+  ExtensionsPackageAppConfigExamples,
 } from '@red-hat-developer-hub/backstage-plugin-marketplace-common';
-import { JsonObject } from '@backstage/types';
 
 import Box from '@mui/material/Box';
 import Grid from '@mui/material/Grid';
@@ -51,103 +45,18 @@ import {
   pluginInstallRouteRef,
   pluginRouteRef,
 } from '../routes';
-import { applyContent, getExampleAsMarkdown } from '../utils';
 
-import {
-  CodeEditorContextProvider,
-  CodeEditor,
-  useCodeEditor,
-} from './CodeEditor';
-import { Markdown } from './Markdown';
+import { CodeEditorContextProvider, useCodeEditor } from './CodeEditor';
 import { usePackage } from '../hooks/usePackage';
+import { CodeEditorCard } from './CodeEditorCard';
+import { TabPanel } from './TabPanel';
 
 interface TabItem {
   label: string;
-  content: string | MarketplacePackageAppConfigExamples[];
+  content: string | ExtensionsPackageAppConfigExamples[];
   key: string;
   others?: { [key: string]: any };
 }
-
-interface TabPanelProps {
-  markdownContent: string | MarketplacePackageAppConfigExamples[];
-  index: number;
-  value: number;
-  others?: { [key: string]: any };
-}
-
-const TabPanel = ({ markdownContent, index, value, others }: TabPanelProps) => {
-  const alertApi = useApi(alertApiRef);
-  const codeEditor = useCodeEditor();
-  if (value !== index) return null;
-
-  const handleApplyContent = (content: string | JsonObject, pkg: string) => {
-    try {
-      const codeEditorContent = codeEditor.getValue();
-      const newContent = applyContent(
-        codeEditorContent || '',
-        pkg,
-        others?.packageName,
-        content,
-      );
-      const selection = codeEditor.getSelection();
-      const position = codeEditor.getPosition();
-      if (newContent) {
-        codeEditor.setValue(newContent);
-        if (selection) {
-          codeEditor.setSelection(selection);
-        }
-        if (position) {
-          codeEditor.setPosition(position);
-        }
-      }
-    } catch (error) {
-      alertApi.post({
-        display: 'transient',
-        severity: 'warning',
-        message: `Could not apply YAML: ${error}`,
-      });
-    }
-  };
-
-  return (
-    <Box
-      role="tabpanel"
-      sx={{ flex: 1, overflow: 'auto', p: 2, scrollbarWidth: 'thin' }}
-    >
-      <Typography component="div">
-        {Array.isArray(markdownContent) ? (
-          markdownContent.map((item, idx) => (
-            <Box key={idx} sx={{ mb: 3 }}>
-              <Typography variant="h5" sx={{ fontWeight: 'bold', mb: 1 }}>
-                {Object.keys(item)[0]}
-              </Typography>
-              {item[Object.keys(item)[0]]?.map?.(ex => (
-                <>
-                  <Typography variant="h6" sx={{ fontWeight: 'bold', mb: 1 }}>
-                    {ex.title}
-                    {ex.content !== 'string' && (
-                      <Button
-                        sx={{ float: 'right' }}
-                        onClick={() =>
-                          handleApplyContent(ex.content, Object.keys(item)[0])
-                        }
-                      >
-                        Apply
-                      </Button>
-                    )}
-                  </Typography>
-                  <Markdown content={getExampleAsMarkdown(ex.content)} />
-                </>
-              ))}
-            </Box>
-          ))
-        ) : (
-          <Markdown content={markdownContent} />
-        )}
-      </Typography>
-    </Box>
-  );
-};
 
 export const MarketplacePackageInstallContent = ({
   pkg,
@@ -188,14 +97,14 @@ export const MarketplacePackageInstallContent = ({
   const navigate = useNavigate();
   const examples = [
     {
-      [`${pkg.metadata.name}`]: pkg?.spec?.appConfigExamples,
+      [`${pkg.metadata.name}`]: pkg.spec?.appConfigExamples,
     },
   ];
   const packageDynamicArtifacts = {
     [`${pkg.metadata.name}`]: pkg.spec?.dynamicArtifact,
   };
   const availableTabs = [
-    examples && {
+    !!Object.values(examples[0])[0] && {
       label: 'Examples',
       content: examples,
       key: 'examples',
@@ -224,34 +133,7 @@ export const MarketplacePackageInstallContent = ({
         spacing={3}
         sx={{ flex: 1, overflow: 'hidden', height: '100%', pb: 1 }}
       >
-        <Grid
-          item
-          xs={12}
-          md={6.5}
-          sx={{ display: 'flex', flexDirection: 'column', height: '100%' }}
-        >
-          <Card
-            sx={{
-              flex: 1,
-              display: 'flex',
-              flexDirection: 'column',
-              overflow: 'hidden',
-              borderRadius: 0,
-            }}
-          >
-            <CardContent
-              sx={{
-                flex: 1,
-                display: 'flex',
-                flexDirection: 'column',
-                overflow: 'auto',
-                scrollbarWidth: 'thin',
-              }}
-            >
-              <CodeEditor defaultLanguage="yaml" onLoaded={onLoaded} />
-            </CardContent>
-          </Card>
-        </Grid>
+        <CodeEditorCard onLoad={onLoaded} />
 
         {showRightCard && (
           <Grid
