+++ workspaces/marketplace/plugins/marketplace/src/components/MarketplacePluginInstallContent.tsx
@@ -22,22 +22,16 @@ import {
   Progress,
   WarningPanel,
 } from '@backstage/core-components';
-import { JsonObject } from '@backstage/types';
 
-import {
-  alertApiRef,
-  useApi,
-  useRouteRef,
-  useRouteRefParams,
-} from '@backstage/core-plugin-api';
+import { useRouteRef, useRouteRefParams } from '@backstage/core-plugin-api';
 
 import yaml from 'yaml';
 import { useNavigate } from 'react-router-dom';
 
 import {
+  ExtensionsPackageAppConfigExamples,
   MarketplacePackage,
   MarketplacePackageSpec,
-  MarketplacePackageAppConfigExamples,
   MarketplacePlugin,
   MarketplacePluginInstallStatus,
 } from '@red-hat-developer-hub/backstage-plugin-marketplace-common';
@@ -63,22 +57,15 @@ import { pluginInstallRouteRef, pluginRouteRef } from '../routes';
 import { usePlugin } from '../hooks/usePlugin';
 import { usePluginPackages } from '../hooks/usePluginPackages';
 import {
-  applyContent,
   DYNAMIC_PLUGIN_CONFIG_YAML,
   EXTENSIONS_CONFIG_YAML,
   ExtensionsStatus,
   getErrorMessage,
-  getExampleAsMarkdown,
   getPluginActionTooltipMessage,
   isPluginInstalled,
 } from '../utils';
 
-import {
-  CodeEditorContextProvider,
-  CodeEditor,
-  useCodeEditor,
-} from './CodeEditor';
-import { Markdown } from './Markdown';
+import { CodeEditorContextProvider, useCodeEditor } from './CodeEditor';
 import {
   InstallationType,
   useInstallationContext,
@@ -91,6 +78,8 @@ import { useNodeEnvironment } from '../hooks/useNodeEnvironment';
 import { useExtensionsConfiguration } from '../hooks/useExtensionsConfiguration';
 import { mapMarketplacePluginInstallStatusToInstallPageButton } from '../labels';
 import { useTheme } from '@mui/material/styles';
+import { CodeEditorCard } from './CodeEditorCard';
+import { TabPanel } from './TabPanel';
 
 const generateCheckboxList = (packages: MarketplacePackage[]) => {
   const hasFrontend = packages.some(
@@ -137,91 +126,10 @@ const CheckboxList = ({ packages }: { packages: MarketplacePackage[] }) => {
 
 interface TabItem {
   label: string;
-  content: string | MarketplacePackageAppConfigExamples[];
+  content: string | ExtensionsPackageAppConfigExamples[];
   key: string;
   others?: { [key: string]: any };
 }
-
-interface TabPanelProps {
-  markdownContent: string | MarketplacePackageAppConfigExamples[];
-  index: number;
-  value: number;
-  others?: { [key: string]: any };
-}
-
-const TabPanel = ({ markdownContent, index, value, others }: TabPanelProps) => {
-  const alertApi = useApi(alertApiRef);
-  const codeEditor = useCodeEditor();
-  if (value !== index) return null;
-
-  const handleApplyContent = (content: string | JsonObject, pkg: string) => {
-    try {
-      const codeEditorContent = codeEditor.getValue();
-      const newContent = applyContent(
-        codeEditorContent || '',
-        pkg,
-        others?.packageNames,
-        content,
-      );
-      const selection = codeEditor.getSelection();
-      const position = codeEditor.getPosition();
-      if (newContent) {
-        codeEditor.setValue(newContent);
-        if (selection) {
-          codeEditor.setSelection(selection);
-        }
-        if (position) {
-          codeEditor.setPosition(position);
-        }
-      }
-    } catch (error) {
-      alertApi.post({
-        display: 'transient',
-        severity: 'warning',
-        message: `Could not apply YAML: ${error}`,
-      });
-    }
-  };
-
-  return (
-    <Box
-      role="tabpanel"
-      sx={{ flex: 1, overflow: 'auto', p: 2, scrollbarWidth: 'thin' }}
-    >
-      <Typography component="div">
-        {Array.isArray(markdownContent) ? (
-          markdownContent.map((item, idx) => (
-            <Box key={idx} sx={{ mb: 3 }}>
-              <Typography variant="h5" sx={{ fontWeight: 'bold', mb: 1 }}>
-                {Object.keys(item)[0]}
-              </Typography>
-              {item[Object.keys(item)[0]].map(ex => (
-                <>
-                  <Typography variant="h6" sx={{ fontWeight: 'bold', mb: 1 }}>
-                    {ex.title}
-                    {ex.content !== 'string' && (
-                      <Button
-                        sx={{ float: 'right' }}
-                        onClick={() =>
-                          handleApplyContent(ex.content, Object.keys(item)[0])
-                        }
-                      >
-                        Apply
-                      </Button>
-                    )}
-                  </Typography>
-                  <Markdown content={getExampleAsMarkdown(ex.content)} />
-                </>
-              ))}
-            </Box>
-          ))
-        ) : (
-          <Markdown content={markdownContent} />
-        )}
-      </Typography>
-    </Box>
-  );
-};
 
 export const MarketplacePluginInstallContent = ({
   plugin,
@@ -486,36 +394,7 @@ export const MarketplacePluginInstallContent = ({
           spacing={3}
           sx={{ flex: 1, overflow: 'hidden', height: '100%', pb: 1 }}
         >
-          {packages.length > 0 && (
-            <Grid
-              item
-              xs={12}
-              md={6.5}
-              sx={{ display: 'flex', flexDirection: 'column', height: '100%' }}
-            >
-              <Card
-                sx={{
-                  flex: 1,
-                  display: 'flex',
-                  flexDirection: 'column',
-                  overflow: 'hidden',
-                  borderRadius: 0,
-                }}
-              >
-                <CardContent
-                  sx={{
-                    flex: 1,
-                    display: 'flex',
-                    flexDirection: 'column',
-                    overflow: 'auto',
-                    scrollbarWidth: 'thin',
-                  }}
-                >
-                  <CodeEditor defaultLanguage="yaml" onLoaded={onLoaded} />
-                </CardContent>
-              </Card>
-            </Grid>
-          )}
+          {packages.length > 0 && <CodeEditorCard onLoad={onLoaded} />}
 
           {showRightCard && (
             <Grid
